refactor: rename voicenotes service module
